+++ src/app/engine/actions.ts
@@ -1,7 +1,31 @@
 "use server";
 
+import { redirect, RedirectType } from "next/navigation";
+import { z } from "zod";
+
+const searchEngines = {
+  duckDuckGo: (query: string) =>
+    `https://duckduckgo.com/?q=${encodeURIComponent(query)}`,
+  google: (query: string) =>
+    `https://www.google.com/search?q=${encodeURIComponent(query)}`,
+  startpage: (query: string) =>
+    `https://www.startpage.com/do/search?q=${encodeURIComponent(query)}`,
+  yandex: (query: string) =>
+    `https://yandex.com/search/?text=${encodeURIComponent(query)}`,
+};
+
+const schema = z.object({
+  search: z.string().min(1, "Search query is required"),
+  engine: z.enum(["duckDuckGo", "google", "startpage", "yandex"]),
+});
+
 export async function performSearch(formData: FormData) {
-  // TODO
-  console.log(formData.get("engine"));
-  console.log(JSON.stringify(Object.fromEntries(formData.entries()), null, 2));
+  const rawData = Object.fromEntries(formData.entries());
+  const data = schema.parse(rawData);
+
+  const { search, engine } = data;
+
+  const searchUrl = searchEngines[engine](search);
+
+  redirect(searchUrl);
 }
